move pub packages to pkg/
