+++ internal/login/token.go
@@ -3,8 +3,8 @@ package login
 import (
 	"charm.land/bubbles/v2/textinput"
 	tea "charm.land/bubbletea/v2"
-	"github.com/ayn2op/discordo/form"
 	"github.com/ayn2op/discordo/internal/keyring"
+	"github.com/ayn2op/discordo/pkg/form"
 )
 
 type TokenMsg struct {
