+++ pkg/tabs/model.go
