+++ pkg/tabs/styles.go
