--- tree/model.go
@@ -1,21 +1,0 @@
-package tree
-
-import tea "charm.land/bubbletea/v2"
-
-type Model struct{}
-
-func NewModel() Model {
-	return Model{}
-}
-
-func (m Model) Init() tea.Cmd {
-	return nil
-}
-
-func (m Model) Update(msg tea.Msg) (Model, tea.Cmd) {
-	return m, nil
-}
-
-func (m Model) View() tea.View {
-	return tea.NewView("")
-}
